с откладкой

+++ main/main.cpp
@@ -52,7 +52,7 @@ int main() {
     // Вычисление определителя
     try {
         auto start = high_resolution_clock::now();
-        int determinant = mat.determinant();
+        double determinant = mat.determinant();
         auto end = high_resolution_clock::now();
         duration<double> elapsed = end - start;
         cout << "Определитель матрицы: " << determinant << endl;
